Move DatasetTbl to views folder, change Dataset link position in navbar
