+++ api/mapping/templates/mapping/dataset_list.html
@@ -11,9 +11,7 @@
 
 <div>
     <div id='root'></div>
-    <script type="text/javascript" >
-        var isAdmin = "{{is_admin}}" === "True" ? true : false
-    </script>
+
     <script type="module" src="../../../static/javascript/react/src/datasetsTable.js"></script>
 
 </div>
