+++ api/templates/base.html
@@ -50,10 +50,10 @@
                     <a class="nav-link" aria-current="page" href="{% url 'home' %}">Home</a>
                 </li>
                 <li class="nav-item active">
-                    <a class="nav-link" aria-current="page" href="{% url 'scan-report-list' %}">Scan Reports</a>
+                    <a class="nav-link" aria-current="page" href="{% url 'datasets' %}">Datasets</a>
                 </li>
                 <li class="nav-item active">
-                    <a class="nav-link" aria-current="page" href="{% url 'datasets' %}">Datasets</a>
+                    <a class="nav-link" aria-current="page" href="{% url 'scan-report-list' %}">Scan Reports</a>
                 </li>
                 <!--<li class="nav-item"><a class="nav-link">You are logged as: {{ user.username }}</a></li>-->
                 <li class="nav-item"><a class="nav-link" href="https://co-connect.github.io/docs/" target="_blank">Documentation</a></li>
